订单完

+++ src/main/java/spring/aoyo_hou/pojo/SysUser.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class SysUser {
 
-  private long userId;
-  private long groupId;
-  private long roleId;
+  private Integer userId;
+  private Integer groupId;
+  private Integer roleId;
   private String userName;
   private String userPwd;
   private String userAlias;
@@ -16,28 +16,28 @@ public class SysUser {
   private String userQq;
   private String userEmail;
   private String userIdentity;
-  private long userAddressId;
+  private Integer userAddressId;
   private String userImagePath;
-  private long countyId;
+  private Integer countyId;
   private String userMac;
   private String userImei;
   private String userToken;
-  private long loginFlag;
+  private Integer loginFlag;
   private String loginDesc;
   private java.sql.Timestamp userLastLogin;
   private String loginUser;
-  private long userSource;
+  private Integer userSource;
   private String userCreator;
   private java.sql.Timestamp userCreateTime;
   private java.sql.Timestamp userUpdateTime;
   private String userWxOpenid;
   private String userWxUnionid;
-  private long workStart;
-  private long workEnd;
-  private long childrenAccountId;
-  private long userStatus;
-  private long userFlag;
-  private long userSex;
+  private Integer workStart;
+  private Integer workEnd;
+  private Integer childrenAccountId;
+  private Integer userStatus;
+  private Integer userFlag;
+  private Integer userSex;
   private String userWeixinName;
